flask/requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
flask~=3.0.0
groq~=0.12.0
langchain~=0.3.8
beautifulsoup4~=4.12.3
ipython~=8.27.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
faiss==1.9.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
gunicorn==23.0.0
ttpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.145
llama-cloud==0.1.5
llama-index==0.12.1
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.1
llama-index-embeddings-openai==0.3.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.0
llama-index-llms-openai-like==0.3.0
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-index-tools-brave-search==0.3.0
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.5.14
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
openai==1.54.5
orjson==3.10.12
pandas==2.2.3
pillow==11.0.0
ply==3.11
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pypdf==5.1.0
PyQt5==5.15.10
python-dotenv==1.0.1
regex==2024.11.6
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
tzdata==2024.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.2
